app/requirements.txt
fastapi
uvicorn
prometheus_clientaltgraph @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.10.7
exceptiongroup==1.3.1
future @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.10
iniconfig==2.1.0
jproperties==2.1.2
macholib @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
packaging==25.0
pluggy==1.6.0
pyfakefs==5.9.3
Pygments==2.19.2
pysonar==1.2.1.3951
pytest==8.4.2
pytest-cov==7.0.0
PyYAML==6.0.3
requests==2.32.5
responses==0.25.8
six @ file:///AppleInternal/Library/BuildRoots/4~CGYkugDNMJDPussZJfa8Q7q7wDj7yZMIh_VI6H4/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.2.3
